Contribution guide + code refactor + package updates
