+++ apps/api/.env.example
@@ -1,6 +1,7 @@
-COOLIFY_APP_ID=
-COOLIFY_SECRET_KEY="12341234123412341234123412341234 - 32 long"
-COOLIFY_DATABASE_URL=
+COOLIFY_APP_ID=local-dev
+# 32 bits long secret key
+COOLIFY_SECRET_KEY=12341234123412341234123412341234
+COOLIFY_DATABASE_URL=file:../db/dev.db
 COOLIFY_SENTRY_DSN=
 
 COOLIFY_IS_ON=docker
