+++ apps/api/package.json
@@ -16,26 +16,26 @@
   "dependencies": {
     "@breejs/ts-worker": "2.0.0",
     "@fastify/autoload": "5.2.0",
-    "@fastify/cookie": "8.0.0",
+    "@fastify/cookie": "8.1.0",
     "@fastify/cors": "8.1.0",
     "@fastify/env": "4.1.0",
     "@fastify/jwt": "6.3.2",
     "@fastify/static": "6.5.0",
     "@iarna/toml": "2.2.5",
     "@ladjs/graceful": "3.0.2",
-    "@prisma/client": "3.15.2",
+    "@prisma/client": "4.2.1",
     "axios": "0.27.2",
     "bcryptjs": "2.4.3",
     "bree": "9.1.2",
     "cabin": "9.1.2",
-    "compare-versions": "4.1.4",
+    "compare-versions": "5.0.1",
     "cuid": "2.1.8",
     "dayjs": "1.11.5",
     "dockerode": "3.3.4",
     "dotenv-extended": "2.9.0",
     "execa": "6.1.0",
-    "fastify": "4.5.2",
-    "fastify-plugin": "4.2.0",
+    "fastify": "4.5.3",
+    "fastify-plugin": "4.2.1",
     "generate-password": "1.7.0",
     "got": "12.3.1",
     "is-ip": "5.0.0",
@@ -57,12 +57,12 @@
     "@typescript-eslint/eslint-plugin": "5.35.1",
     "@typescript-eslint/parser": "5.35.1",
     "esbuild": "0.15.5",
-    "eslint": "8.22.0",
+    "eslint": "8.23.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-prettier": "4.2.1",
     "nodemon": "2.0.19",
     "prettier": "2.7.1",
-    "prisma": "3.15.2",
+    "prisma": "4.2.1",
     "rimraf": "3.0.2",
     "tsconfig-paths": "4.1.0",
     "typescript": "4.7.4"
