+++ apps/api/src/lib/common.ts
@@ -15,9 +15,11 @@ import sshConfig from 'ssh-config'
 
 import { checkContainer, removeContainer } from './docker';
 import { day } from './dayjs';
-import * as serviceFields from './serviceFields'
+import * as serviceFields from './services/serviceFields'
 import { saveBuildLog } from './buildPacks/common';
 import { scheduler } from './scheduler';
+import { supportedServiceTypesAndVersions } from './services/supportedVersions';
+import { includeServices } from './services/common';
 
 export const version = '3.8.5';
 export const isDev = process.env.NODE_ENV === 'development';
@@ -66,25 +68,6 @@ const otherTraefikEndpoint = isDev
 	? `${getAPIUrl()}/webhooks/traefik/other.json`
 	: 'http://coolify:3000/webhooks/traefik/other.json';
 
-
-export const include: any = {
-	destinationDocker: true,
-	persistentStorage: true,
-	serviceSecret: true,
-	minio: true,
-	plausibleAnalytics: true,
-	vscodeserver: true,
-	wordpress: true,
-	ghost: true,
-	meiliSearch: true,
-	umami: true,
-	hasura: true,
-	fider: true,
-	moodle: true,
-	appwrite: true,
-	glitchTip: true,
-	searxng: true
-};
 
 export const uniqueName = (): string => uniqueNamesGenerator(customConfig);
 export const asyncExecShell = util.promisify(exec);
@@ -200,199 +183,7 @@ export const encrypt = (text: string) => {
 	}
 };
 
-export const supportedServiceTypesAndVersions = [
-	{
-		name: 'plausibleanalytics',
-		fancyName: 'Plausible Analytics',
-		baseImage: 'plausible/analytics',
-		images: ['bitnami/postgresql:13.2.0', 'yandex/clickhouse-server:21.3.2.5'],
-		versions: ['latest', 'stable'],
-		recommendedVersion: 'stable',
-		ports: {
-			main: 8000
-		}
-	},
-	{
-		name: 'nocodb',
-		fancyName: 'NocoDB',
-		baseImage: 'nocodb/nocodb',
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 8080
-		}
-	},
-	{
-		name: 'minio',
-		fancyName: 'MinIO',
-		baseImage: 'minio/minio',
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 9001
-		}
-	},
-	{
-		name: 'vscodeserver',
-		fancyName: 'VSCode Server',
-		baseImage: 'codercom/code-server',
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 8080
-		}
-	},
-	{
-		name: 'wordpress',
-		fancyName: 'Wordpress',
-		baseImage: 'wordpress',
-		images: ['bitnami/mysql:5.7'],
-		versions: ['latest', 'php8.1', 'php8.0', 'php7.4', 'php7.3'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 80
-		}
-	},
-	{
-		name: 'vaultwarden',
-		fancyName: 'Vaultwarden',
-		baseImage: 'vaultwarden/server',
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 80
-		}
-	},
-	{
-		name: 'languagetool',
-		fancyName: 'LanguageTool',
-		baseImage: 'silviof/docker-languagetool',
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 8010
-		}
-	},
-	{
-		name: 'n8n',
-		fancyName: 'n8n',
-		baseImage: 'n8nio/n8n',
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 5678
-		}
-	},
-	{
-		name: 'uptimekuma',
-		fancyName: 'Uptime Kuma',
-		baseImage: 'louislam/uptime-kuma',
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 3001
-		}
-	},
-	{
-		name: 'ghost',
-		fancyName: 'Ghost',
-		baseImage: 'bitnami/ghost',
-		images: ['bitnami/mariadb'],
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 2368
-		}
-	},
-	{
-		name: 'meilisearch',
-		fancyName: 'Meilisearch',
-		baseImage: 'getmeili/meilisearch',
-		images: [],
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 7700
-		}
-	},
-	{
-		name: 'umami',
-		fancyName: 'Umami',
-		baseImage: 'ghcr.io/mikecao/umami',
-		images: ['postgres:12-alpine'],
-		versions: ['postgresql-latest'],
-		recommendedVersion: 'postgresql-latest',
-		ports: {
-			main: 3000
-		}
-	},
-	{
-		name: 'hasura',
-		fancyName: 'Hasura',
-		baseImage: 'hasura/graphql-engine',
-		images: ['postgres:12-alpine'],
-		versions: ['latest', 'v2.10.0', 'v2.5.1'],
-		recommendedVersion: 'v2.10.0',
-		ports: {
-			main: 8080
-		}
-	},
-	{
-		name: 'fider',
-		fancyName: 'Fider',
-		baseImage: 'getfider/fider',
-		images: ['postgres:12-alpine'],
-		versions: ['stable'],
-		recommendedVersion: 'stable',
-		ports: {
-			main: 3000
-		}
-	},
-	{
-		name: 'appwrite',
-		fancyName: 'Appwrite',
-		baseImage: 'appwrite/appwrite',
-		images: ['mariadb:10.7', 'redis:6.2-alpine', 'appwrite/telegraf:1.4.0'],
-		versions: ['latest', '0.15.3'],
-		recommendedVersion: '0.15.3',
-		ports: {
-			main: 80
-		}
-	},
-	// {
-	//     name: 'moodle',
-	//     fancyName: 'Moodle',
-	//     baseImage: 'bitnami/moodle',
-	//     images: [],
-	//     versions: ['latest', 'v4.0.2'],
-	//     recommendedVersion: 'latest',
-	//     ports: {
-	//         main: 8080
-	//     }
-	// }
-	{
-		name: 'glitchTip',
-		fancyName: 'GlitchTip',
-		baseImage: 'glitchtip/glitchtip',
-		images: ['postgres:14-alpine', 'redis:7-alpine'],
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 8000
-		}
-	},
-	{
-		name: 'searxng',
-		fancyName: 'SearXNG',
-		baseImage: 'searxng/searxng',
-		images: [],
-		versions: ['latest'],
-		recommendedVersion: 'latest',
-		ports: {
-			main: 8080
-		}
-	},
-];
+
 
 export async function checkDoubleBranch(branch: string, projectId: number): Promise<boolean> {
 	const applications = await prisma.application.findMany({ where: { branch, projectId } });
@@ -570,7 +361,7 @@ export function generateTimestamp(): string {
 
 export async function listServicesWithIncludes(): Promise<any> {
 	return await prisma.service.findMany({
-		include,
+		include: includeServices,
 		orderBy: { createdAt: 'desc' }
 	});
 }
@@ -1497,311 +1288,6 @@ export function getServiceImages(type: string): string[] {
 		return found.images;
 	}
 	return [];
-}
-
-export async function configureServiceType({
-	id,
-	type
-}: {
-	id: string;
-	type: string;
-}): Promise<void> {
-	if (type === 'plausibleanalytics') {
-		const password = encrypt(generatePassword({}));
-		const postgresqlUser = cuid();
-		const postgresqlPassword = encrypt(generatePassword({}));
-		const postgresqlDatabase = 'plausibleanalytics';
-		const secretKeyBase = encrypt(generatePassword({ length: 64 }));
-
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				plausibleAnalytics: {
-					create: {
-						postgresqlDatabase,
-						postgresqlUser,
-						postgresqlPassword,
-						password,
-						secretKeyBase
-					}
-				}
-			}
-		});
-	} else if (type === 'nocodb') {
-		await prisma.service.update({
-			where: { id },
-			data: { type }
-		});
-	} else if (type === 'minio') {
-		const rootUser = cuid();
-		const rootUserPassword = encrypt(generatePassword({}));
-		await prisma.service.update({
-			where: { id },
-			data: { type, minio: { create: { rootUser, rootUserPassword } } }
-		});
-	} else if (type === 'vscodeserver') {
-		const password = encrypt(generatePassword({}));
-		await prisma.service.update({
-			where: { id },
-			data: { type, vscodeserver: { create: { password } } }
-		});
-	} else if (type === 'wordpress') {
-		const mysqlUser = cuid();
-		const mysqlPassword = encrypt(generatePassword({}));
-		const mysqlRootUser = cuid();
-		const mysqlRootUserPassword = encrypt(generatePassword({}));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				wordpress: { create: { mysqlPassword, mysqlRootUserPassword, mysqlRootUser, mysqlUser } }
-			}
-		});
-	} else if (type === 'vaultwarden') {
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type
-			}
-		});
-	} else if (type === 'languagetool') {
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type
-			}
-		});
-	} else if (type === 'n8n') {
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type
-			}
-		});
-	} else if (type === 'uptimekuma') {
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type
-			}
-		});
-	} else if (type === 'ghost') {
-		const defaultEmail = `${cuid()}@example.com`;
-		const defaultPassword = encrypt(generatePassword({}));
-		const mariadbUser = cuid();
-		const mariadbPassword = encrypt(generatePassword({}));
-		const mariadbRootUser = cuid();
-		const mariadbRootUserPassword = encrypt(generatePassword({}));
-
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				ghost: {
-					create: {
-						defaultEmail,
-						defaultPassword,
-						mariadbUser,
-						mariadbPassword,
-						mariadbRootUser,
-						mariadbRootUserPassword
-					}
-				}
-			}
-		});
-	} else if (type === 'meilisearch') {
-		const masterKey = encrypt(generatePassword({ length: 32 }));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				meiliSearch: { create: { masterKey } }
-			}
-		});
-	} else if (type === 'umami') {
-		const umamiAdminPassword = encrypt(generatePassword({}));
-		const postgresqlUser = cuid();
-		const postgresqlPassword = encrypt(generatePassword({}));
-		const postgresqlDatabase = 'umami';
-		const hashSalt = encrypt(generatePassword({ length: 64 }));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				umami: {
-					create: {
-						umamiAdminPassword,
-						postgresqlDatabase,
-						postgresqlPassword,
-						postgresqlUser,
-						hashSalt
-					}
-				}
-			}
-		});
-	} else if (type === 'hasura') {
-		const postgresqlUser = cuid();
-		const postgresqlPassword = encrypt(generatePassword({}));
-		const postgresqlDatabase = 'hasura';
-		const graphQLAdminPassword = encrypt(generatePassword({}));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				hasura: {
-					create: {
-						postgresqlDatabase,
-						postgresqlPassword,
-						postgresqlUser,
-						graphQLAdminPassword
-					}
-				}
-			}
-		});
-	} else if (type === 'fider') {
-		const postgresqlUser = cuid();
-		const postgresqlPassword = encrypt(generatePassword({}));
-		const postgresqlDatabase = 'fider';
-		const jwtSecret = encrypt(generatePassword({ length: 64, symbols: true }));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				fider: {
-					create: {
-						postgresqlDatabase,
-						postgresqlPassword,
-						postgresqlUser,
-						jwtSecret
-					}
-				}
-			}
-		});
-	} else if (type === 'moodle') {
-		const defaultUsername = cuid();
-		const defaultPassword = encrypt(generatePassword({}));
-		const defaultEmail = `${cuid()} @example.com`;
-		const mariadbUser = cuid();
-		const mariadbPassword = encrypt(generatePassword({}));
-		const mariadbDatabase = 'moodle_db';
-		const mariadbRootUser = cuid();
-		const mariadbRootUserPassword = encrypt(generatePassword({}));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				moodle: {
-					create: {
-						defaultUsername,
-						defaultPassword,
-						defaultEmail,
-						mariadbUser,
-						mariadbPassword,
-						mariadbDatabase,
-						mariadbRootUser,
-						mariadbRootUserPassword
-					}
-				}
-			}
-		});
-	} else if (type === 'appwrite') {
-		const opensslKeyV1 = encrypt(generatePassword({}));
-		const executorSecret = encrypt(generatePassword({}));
-		const redisPassword = encrypt(generatePassword({}));
-		const mariadbHost = `${id}-mariadb`
-		const mariadbUser = cuid();
-		const mariadbPassword = encrypt(generatePassword({}));
-		const mariadbDatabase = 'appwrite';
-		const mariadbRootUser = cuid();
-		const mariadbRootUserPassword = encrypt(generatePassword({}));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				appwrite: {
-					create: {
-						opensslKeyV1,
-						executorSecret,
-						redisPassword,
-						mariadbHost,
-						mariadbUser,
-						mariadbPassword,
-						mariadbDatabase,
-						mariadbRootUser,
-						mariadbRootUserPassword
-					}
-				}
-			}
-		});
-	} else if (type === 'glitchTip') {
-		const defaultUsername = cuid();
-		const defaultEmail = `${defaultUsername}@example.com`;
-		const defaultPassword = encrypt(generatePassword({}));
-		const postgresqlUser = cuid();
-		const postgresqlPassword = encrypt(generatePassword({}));
-		const postgresqlDatabase = 'glitchTip';
-		const secretKeyBase = encrypt(generatePassword({ length: 64 }));
-
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				glitchTip: {
-					create: {
-						postgresqlDatabase,
-						postgresqlUser,
-						postgresqlPassword,
-						secretKeyBase,
-						defaultEmail,
-						defaultUsername,
-						defaultPassword,
-					}
-				}
-			}
-		});
-	} else if (type === 'searxng') {
-		const secretKey = encrypt(generatePassword({ length: 32, isHex: true }))
-		const redisPassword = encrypt(generatePassword({}));
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type,
-				searxng: {
-					create: {
-						secretKey,
-						redisPassword,
-					}
-				}
-			}
-		});
-	} else {
-		await prisma.service.update({
-			where: { id },
-			data: {
-				type
-			}
-		});
-	}
-}
-
-export async function removeService({ id }: { id: string }): Promise<void> {
-	await prisma.serviceSecret.deleteMany({ where: { serviceId: id } });
-	await prisma.servicePersistentStorage.deleteMany({ where: { serviceId: id } });
-	await prisma.meiliSearch.deleteMany({ where: { serviceId: id } });
-	await prisma.fider.deleteMany({ where: { serviceId: id } });
-	await prisma.ghost.deleteMany({ where: { serviceId: id } });
-	await prisma.umami.deleteMany({ where: { serviceId: id } });
-	await prisma.hasura.deleteMany({ where: { serviceId: id } });
-	await prisma.plausibleAnalytics.deleteMany({ where: { serviceId: id } });
-	await prisma.minio.deleteMany({ where: { serviceId: id } });
-	await prisma.vscodeserver.deleteMany({ where: { serviceId: id } });
-	await prisma.wordpress.deleteMany({ where: { serviceId: id } });
-	await prisma.glitchTip.deleteMany({ where: { serviceId: id } });
-	await prisma.moodle.deleteMany({ where: { serviceId: id } });
-	await prisma.appwrite.deleteMany({ where: { serviceId: id } });
-	await prisma.searxng.deleteMany({ where: { serviceId: id } });
-	await prisma.service.delete({ where: { id } });
 }
 
 export function saveUpdateableFields(type: string, data: any) {
