+++ apps/ui/src/routes/__layout.svelte
@@ -65,11 +65,13 @@
 
 <script lang="ts">
 	export let baseSettings: any;
+	export let supportedServiceTypesAndVersions: any;
 	$appSession.ipv4 = baseSettings.ipv4;
 	$appSession.ipv6 = baseSettings.ipv6;
 	$appSession.version = baseSettings.version;
 	$appSession.whiteLabeled = baseSettings.whiteLabeled;
 	$appSession.whiteLabeledDetails.icon = baseSettings.whiteLabeledIcon;
+	$appSession.supportedServiceTypesAndVersions = supportedServiceTypesAndVersions
 
 	export let userId: string;
 	export let teamId: string;
